fix(episode-flow): improve check-in hub semantics, a11y feedback, and cross-platform history/food diary UX consistency
