+++ apps/web/src/components/episode-flow/SymptomPromptFlow.tsx
@@ -17,10 +17,12 @@ import type {
 } from '@abstrack/types';
 import type { EpisodeRow } from '@abstrack/types';
 import {
+  compareEpisodeSymptomRowsForHistory,
   computeSymptomResumePlacement,
   createDefaultSymptomPromptAnswer,
   createInitialSymptomPromptSession,
   episodeSymptomRowsToAnswersMapForOpenPass,
+  formatEpisodeSymptomHistoryDetail,
   formatEpisodeDurationSimple,
   symptomPromptAnswerHasValue,
 } from '@abstrack/types';
@@ -77,48 +79,6 @@ function clampIndex(index: number, length: number): number {
 /** Queue key so the same preset symptom id does not serialize writes across episodes. */
 function lineWriteQueueKey(episodeId: string, presetSymptomId: string): string {
   return `${episodeId}:${presetSymptomId}`;
-}
-
-function compareSymptomRowsForHistory(
-  a: EpisodeSymptomRow,
-  b: EpisodeSymptomRow,
-): number {
-  const aMs = Date.parse(a.created_at);
-  const bMs = Date.parse(b.created_at);
-  const aValid = Number.isFinite(aMs);
-  const bValid = Number.isFinite(bMs);
-  if (aValid && bValid) {
-    const byTime = aMs - bMs;
-    if (byTime !== 0) {
-      return byTime;
-    }
-  } else {
-    const byText = a.created_at.localeCompare(b.created_at);
-    if (byText !== 0) {
-      return byText;
-    }
-  }
-  return a.id.localeCompare(b.id);
-}
-
-function symptomHistoryDetail(row: EpisodeSymptomRow): string {
-  if (row.response_type === 'yes_no' && row.response_boolean != null) {
-    return row.response_boolean ? 'Yes' : 'No';
-  }
-  if (row.response_type === 'severity_scale' && row.response_severity != null) {
-    return `Severity ${row.response_severity}`;
-  }
-  if (row.response_type === 'free_text') {
-    const text = row.response_text?.trim() ?? '';
-    return text.length > 0 ? text : '—';
-  }
-  if (row.response_type === 'photo') {
-    return 'Photo';
-  }
-  if (row.response_type === 'video') {
-    return 'Video';
-  }
-  return '—';
 }
 
 /**
@@ -591,7 +551,7 @@ export function SymptomPromptFlow({
       : {};
     if (fromServer.ok) {
       setSymptomHistory(
-        fromServer.data.slice().sort(compareSymptomRowsForHistory),
+        fromServer.data.slice().sort(compareEpisodeSymptomRowsForHistory),
       );
     } else {
       setSymptomHistory([]);
@@ -1020,7 +980,7 @@ export function SymptomPromptFlow({
                     <EpisodeLocaleInstant iso={row.created_at} />
                     {' — '}
                   </span>
-                  {row.symptom_name}: {symptomHistoryDetail(row)}
+                  {row.symptom_name}: {formatEpisodeSymptomHistoryDetail(row)}
                 </li>
               ))}
             </ol>
